# scheduler: add CustomLimiter interface for extensible resource limiting in ElasticQuota plugin