### Ⅰ. Describe what this PR does

<!--
- Summarize your change (**mandatory**)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)
-->

This PR introduces a new feature to the ElasticQuota scheduler plugin by adding a `CustomLimiter` interface. This interface allows users to define custom logic for limiting resource usage within quotas, providing greater flexibility and control over resource management.

**CustomLimiter Interface**  
   The `CustomLimiter` interface provides the following methods:  
   - `GetKey()`: Returns the unique identifier for the custom limiter.
   - `GetLimitConf(quota *v1alpha1.ElasticQuota) (*CustomLimitConf, error)`: Parses the custom limit configuration for a given quota. Returns nil if no limit is configured for this quota, or non-nil CustomLimitConf if custom limiter is configured.
   - `CalculatePodUsedDelta(quotaInfo *QuotaInfo, newPod, oldPod *corev1.Pod, quotaChainSharedState *CustomLimiterState) corev1.ResourceList`: Calculates the delta of used resources for a pod update, returns nil if no limit is configured for this quota, returns non-nil resource if current custom limiter is configured for this quota.
   - `Check(quotaInfo *QuotaInfo, podRequest corev1.ResourceList, quotaChainSharedState *CustomLimiterState) error`: Validates whether a pod's request complies with the custom limit, the check process should start from the leaf quota and proceed to parents.

### How to use
1. Create a custom limiter by implementing the CustomLimiter interface, such as MockCustomLimiter, which defines resource limits for pods with specific labels. Then register the CustomLimiter factory as following:
```
RegisterCustomLimiterFactory("mockFactory", func(key, args string) (CustomLimiter, error){
	return &MockCustomLimiter{
		key:      key,
		selector: parseFromArgs(args),
	}
})
```

2. Add your enabled custom limiters to the scheduler configuration
```
data:
  koord-scheduler-config: |
    apiVersion: kubescheduler.config.k8s.io/v1
    kind: KubeSchedulerConfiguration
    profiles:
      - pluginConfig:
        - name: ElasticQuota
          args:
            apiVersion: kubescheduler.config.k8s.io/v1
            kind: ElasticQuotaArgs
            customLimiters:
              driver:
              	factoryKey: mockFactory
              	factoryArgs: '{"labelSelectors":"role=driver"}'
              spot:
              	factoryKey: mockFactory
              	factoryArgs: '{"labelSelectors":"type=spot"}'
```

3. Set resource limits by adding specified annotations to your ElasticQuota resources:
```
apiVersion: scheduling.sigs.k8s.io/v1alpha1
kind: ElasticQuota
metadata:
  annotations:
    custom-driver-limit-conf: '{"cpu":"10","memory":"20Gi"}'  # limit conf for driver pods
```

4. The actual limit and current used resources will be shown in quota annotations
```
apiVersion: scheduling.sigs.k8s.io/v1alpha1
kind: ElasticQuota
metadata:
  labels:
    quota.scheduling.koordinator.sh/parent: 1
  annotations:
    custom-driver-limit-conf: '{"cpu":"10","memory":"20Gi"}'  # limit conf for driver pods
    quota.scheduling.koordinator.sh/custom-driver-limit: '{"cpu":"10","memory":"20Gi"}'  # actual limit
    quota.scheduling.koordinator.sh/custom-driver-used: '{"cpu":"1","memory":"4Gi"}'	 # currernt used resources
  namespace: default
  name: 12
spec:
  max:
    cpu: "100"
    memory: 200Gi
  min:
    cpu: "10"
    memory: 20Gi
status:
  used:
    cpu: "0"
    memory: "0"
```

### Rules
1. Quota Updates
If a specified custom-limit is configured for a quota, the custom-limiter should be marked as enabled for all the descendants and all the parents (exclude root quota) of this quota. This facilitates quick checks for custom limit restrictions during pod scheduling and fast updates to custom-used resources when pods change. So that the sub-tree of the updated quota could be rebuilt when it changes.
For example, when quota 11 is updated with custom-limit configuration of the custom-limiter named "driver", custom-limiter "driver" should be marked as enabled for quota 121, 122, 12, 1.
```
quota 1      		custom-used: {"driver": {}}
  |- quota 11
  	 |- quota 111
  	 |- quota 112
  |- quota 12      	custom-used: {"driver": {}}	custom-limits: {"driver":{"limit":{cpu:10, memory:10}}} 
  	 |- quota 121  	custom-used: {"driver": {}}
  	 |- quota 122  	custom-used: {"driver": {}}
quota 2
  |- quota 21
  	 |- quota 211
  	 |- quota 212
  |- quota 22
  	 |- quota 221
  	 |- quota 222
```

2. Pod Updates
The custom-used-delta resource for a pod update should be calculated from lowest quota to highest quota, returned if it's not nil, so that the same custom-used-delta resource will be applied to the entire quota chain to keep consistency.
For example, when a matched pod is added into quota 121, the calculation process may be as following:
   -  skip for quota 121 (according to the implementation of CustomLimiter interface, for this case, the calculation process needs the arguments configured on quota 12 to know whether the pod is matched)
   -  get not-nil custom-used-delta resource (e.g. <cpu:1, memory:1>) for quota 12 and then the calculation process will be broken. 
   -  at last, the same custom-used-delta resource will be applied to the entire quota chain(quota 121, 12, 1).
```
quota 1			custom-used: {"driver": {cpu:1, memory:1}}
  |- quota 11
  	 |- quota 111
  	 |- quota 112
  |- quota 12      	custom-used: {"driver": {cpu:1, memory:1}}	  custom-limits: {"driver":{"limit":{cpu:10, memory:10}}}
  	 |- quota 121  	custom-used: {"driver": {cpu:1, memory:1}}
  	 |- quota 122  	custom-used: {"driver": {}}
quota 2
  |- quota 21
  	 |- quota 211
  	 |- quota 212
  |- quota 22
  	 |- quota 221
  	 |- quota 222
```

3. Pod Validation
The scheduler only needs to check quotas that have configured custom-limits in the entire quota chain. Quotas without custom-limits will be quickly skipped.
For the quota tree shown above:
If a pod is submitted to quota 211, no custom-limit validation should be triggered since there are no enabled custom-limiters for quota 211.
If a pod is submitted to quota 122, the custom-limit validation will be skipped for quota 122 and checked for quota 12.
This ensures that the scheduler efficiently skips unnecessary checks, improving performance.



### Ⅱ. Does this pull request fix one issue?

<!--If so, add "fixes #xxxx" below in the next line, for example, fixes #15. Otherwise, add "NONE" -->

### Ⅲ. Describe how to verify it

### Ⅳ. Special notes for reviews

### V. Checklist

- [x] I have written necessary docs and comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `make test`