@TaoYang526 2025-04-01
Key Updates:
1. Create CustomLimitersManager to centrally manage the logic related to custom limiters, achieving functional cohesion for easier maintenance.
2. Optimize the performance for CustomLimiters. Maintain information about enabled custom limiters for each quota, triggering automatic refreshes upon quota updates. This facilitates quick checks for custom limit restrictions during pod scheduling and fast updates to custom-used resources when pods change.
3. Add benchmark tests to conduct performance comparison for processes such as quota updates, pod scheduling checks, and resource deduction during pod changes.
   - Quota updates may trigger the reconstruction of a subtree of quotas. The benchmark test results indicate that reconstructing a subtree with 12,000 leaf quotas and 48,000 pods takes approximately 237ms.
   - When pods change, quotas need to update used and custom-used resources in real-time. The benchmark test results show that for a 4-level quota structure, enabling 3 custom-limiters compared to not enabling any custom-limiters increases the processing time by a few microseconds (21us -> 26.4us).
   - During pod scheduling, the scheduler needs to check resource limits (including enabled custom-limiters) for each level of the quota chain. The benchmark test results indicate that for a 4-level quota structure, enabling 3 custom-limiters compared to not enabling any custom-limiters increases the scheduling check processing time by a few microseconds (4.2us -> 6.6us).