@TaoYang526 2025-04-21
/close
This PR has been replaced by PR #2415 since the `QuotaHookPlugin` interface is more generic and has minimal impact on the existing codebase, as concluded from offline discussions.